Projects síða sýnir nú lista af projects sem user á eða er team member af
