--- src/main/java/is/hi/hbv501g/group_project/home/HomeController.java
@@ -1,18 +1,0 @@
-package is.hi.hbv501g.group_project.home;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.servlet.ModelAndView;
-
-@RestController
-public class HomeController {
-
-    @RequestMapping(value = {""}, method = RequestMethod.GET)
-    public ModelAndView home() {
-        ModelAndView modelAndView = new ModelAndView();
-        modelAndView.setViewName("home"); // resources/template/home.html
-        return modelAndView;
-    }
-}
