+++ src/main/java/is/hi/hbv501g/group_project/project/ProjectController.java
@@ -5,6 +5,7 @@ import is.hi.hbv501g.group_project.registration.RegistrationRequest;
 import lombok.AllArgsConstructor;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.ui.Model;
 import org.springframework.ui.ModelMap;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -16,7 +17,17 @@ import org.springframework.web.servlet.ModelAndView;
 @AllArgsConstructor
 public class ProjectController {
 
-    private final AddProjectService addProjectService;
+    private final ProjectService projectService;
+
+    @RequestMapping(value = {""}, method = RequestMethod.GET)
+    public ModelAndView home() {
+        ModelAndView modelAndView = new ModelAndView();
+        modelAndView.setViewName("home"); // resources/template/home.html
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        AppUser user = (AppUser) authentication.getPrincipal();
+        modelAndView.addObject("projects",projectService.findByUser(user));
+        return modelAndView;
+    }
 
     @RequestMapping(value = {"/addProject"}, method = RequestMethod.GET)
     public ModelAndView addProject(){
@@ -39,7 +50,7 @@ public class ProjectController {
             modelAndView.addObject("successMessage", "Please add correct details!");
             modelMap.addAttribute("bindingResult", bindingResult);
         }else {
-            addProjectService.addProject(request, user);
+            projectService.addProject(request, user);
             modelAndView.addObject("successMessage", "Project added!");
         }
         modelAndView.addObject("project", new Project());
