+++ src/main/java/is/hi/hbv501g/group_project/project/ProjectMembers.java
@@ -7,7 +7,6 @@ import javax.persistence.*;
 public class ProjectMembers {
 
     @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long projectId;
     private long userId;
     private String authorities;
