+++ src/main/java/is/hi/hbv501g/group_project/project/ProjectMembersRepository.java
@@ -1,11 +1,15 @@
 package is.hi.hbv501g.group_project.project;
 
+import is.hi.hbv501g.group_project.appuser.AppUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
 
 @Repository
 @Transactional(readOnly = true)
 public interface ProjectMembersRepository
         extends JpaRepository<ProjectMembers, Long> {
+    List<ProjectMembers> findByUserId(long user);
 }
