+++ src/main/java/is/hi/hbv501g/group_project/project/ProjectRepository.java
@@ -4,9 +4,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.List;
+
 
 @Repository
 @Transactional(readOnly = true)
 public interface ProjectRepository
         extends JpaRepository<Project, Long> {
+    List<Project> findByIdIn(List<Long> projectIdList);
 }
